backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
redis==5.0.1
websockets==12.0
python-dotenv==1.0.0
aiofiles==23.2.1
ffmpeg-python==0.2.0
